refactor(api): extract the OAuth-agent resolver from context (cleanup 4/5)
The OAuth access-token -> scoped-agent resolution (opaque grants + RFC 8707
JWTs) was ~90 lines inside buildContext and the sole jose/JWKS consumer there.
Move it to lib/oauth-agent.ts as makeOauthAgent({ meta, auth, baseUrl,
provisionPersonal }) -> { oauthAgent }; agentFor dispatches to it exactly as
before. The rest of context stays one cohesive closure (the identity/workspace/
authz helpers are coupled through per-request WeakMap caches). Drops context.ts
back under the file-size line and quarantines jose. No behavior change.
